backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
fastapi==0.115.11
h11==0.14.0
idna==3.10
ifcopenshell==0.8.1.post1
isodate==0.7.2
lark==1.2.2
numpy==2.0.2
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
python-multipartpymongo
